Document placement of config file, commands for startup
Config file needs to be put in /etc/... folder before being
recognized, or the scripts will silently fail to read it and
throw errors on connection attempts.

Also add a blank like to ensure the two scripts are shown as
two separate command lines in the webpage presentation of
the README.md file, otherwise it looks like a single commandline.